Updated the _prepareCollection() method
Fixes a fatal error that occurs when products have options in the cart. The admin page tries to load the item collection twice but the second time has fake product ID's from the product options that throws a fatal error in the magento/module-quote/Model/ResourceModel/Quote/Item/Collection.php::removeItemsWithAbsentProducts() method when it tries to remove them since it cannot find the quote item by the fake product ID key. This skips that step if the collection has already been loaded previously so we can just display the already loaded items on the grid.